cambios en la estructura de types, mejora en Navigation
