+++ components/MainScreen.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { View, StyleSheet, TouchableOpacity, Text } from "react-native";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
-import { RootStackParamList } from "../types";
+import { RootStackParamList } from "../types/types";
 
 type MainScreenProps = NativeStackScreenProps<RootStackParamList, "Main">;
